[MM-19] Separate Expenses and Income (#55)
## Why?

For easier viewing, the transaction view needed to be tabbed between
'Income', 'All', and 'Expense' transactions. The table footer should
also be sticky to preserve that information. The transactions also
needed to be paginated.

Note: there is an issue with the table header that needs to be addressed
in the future.

## What Changed

* [x] Create a session variable to handle the tabbed views
* [x] Make the form pre-fill the transaction type with whichever tab is
currently selected (income for all selected)
* [x] Make the table footer sticky
* [x] Fix the navbar fixed position
* [x] Paginate the transactions

## Health Checks

These can be done by entering `rake` in the terminal

* [x] rspec
* [x] rubocop
* [x] audit

## Screenshots

<img width="1034" alt="Screenshot 2025-03-08 at 3 59 42 PM"
src="https://github.com/user-attachments/assets/baabcd58-e386-41ee-8605-d665aa512cf8"
/>